requirements.txt
flask
werkzeug
google-cloud
google-cloud-storage